[refac]: moved translator to `utils/docs/helpers`
